requirements.txt
# Data manipulation and analysis
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.12.2

scikit-learn==1.3.0
tensorflow==2.13.0

# Utilities
Pillow==10.0.0
joblib==1.3.2

statsmodels==0.14.0
scipy==1.11.3

gradio